+++ fe2o3-amqp-ws/Changelog.md
@@ -1,5 +1,9 @@
 # fe2o3-amqp-ws
 
+## 0.10.0
+
+1. Unified versioning with other `fe2o3-amqp` crates
+
 ## 0.9.0
 
 1. Unified versioning with other `fe2o3-amqp` crates