+++ fe2o3-amqp/Changelog.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.10.0
+
+1. Changed `RecvError::MessageDecodeError` to `RecvError::MessageDecode(MessageDecodeError)`.
+2. Added `MessageDecodeError` which carries the source `serde_amqp::Error` when deserializing the
+   message body fails and the delivery info that can be used to dispose the message.
+
 ## 0.9.8
 
 1. Fixed bug with reattaching session with sender.